+++ CHANGELOG.md
@@ -48,6 +48,7 @@ Core - Extensions
 - `Collection.allMatch(_:)` is renamed to `Collection.allSatisfy(_:)`
 - `Collection.containsAnyItem(fromCollection:)` method is removed
 - `KeyPathInitializableEnumeration` is replaced with an extension
+- `View.stableRefreshable(action:)` is added, that ensures `View` stability during a refresh
 
 Core - Global Functions